Deals with true/false issues. poly_bool() should be used for conversion, as it is distributive AND peeks into strings to look for "true" or "false" et al. istrue()/isfalse() should be used for checking Python like truthiness.
Bug fix: error when using strings in blocks. Additionally, all values
are not "CONST" type because we never looked at the name, just pull
the value. As the "deep copy" code has been pulled, we need to see if
CONSTs that are non ordinals should be cloned at the time they are
evaluated.