Bug-fix: math.FormatJson() causes error
All json.dumps() need to have a "default=" arg to keep them
out of trouble. Right now just using str, but it should be something
smarter in the future. See the math.random values: these should
return numbers, but the dynamic value class is not smart enough
to do that, nor are the wrapped objects.